Reorganize unit tests in liveblocks-client
