+++ packages/liveblocks-client/src/__tests__/LiveList.test.ts
@@ -1,9 +1,14 @@
+import { LiveList } from "../LiveList";
+import { LiveMap } from "../LiveMap";
+import { LiveObject } from "../LiveObject";
+import type { IdTuple, SerializedCrdt } from "../types";
+import { CrdtType, OpCode, WebsocketCloseCodes } from "../types";
 import {
   listUpdate,
   listUpdateDelete,
   listUpdateInsert,
   listUpdateMove,
-} from "../test/updatesUtils";
+} from "./_updatesUtils";
 import {
   createSerializedList,
   createSerializedObject,
@@ -17,12 +22,7 @@ import {
   reconnect,
   SECOND_POSITION,
   THIRD_POSITION,
-} from "../test/utils";
-import { LiveList } from "./LiveList";
-import { LiveMap } from "./LiveMap";
-import { LiveObject } from "./LiveObject";
-import type { IdTuple, SerializedCrdt } from "./types";
-import { CrdtType, OpCode, WebsocketCloseCodes } from "./types";
+} from "./_utils";
 
 describe("LiveList", () => {
   describe("not attached", () => {
