+++ packages/liveblocks-client/src/__tests__/position.test.ts
@@ -1,4 +1,4 @@
-import { makePosition, max, min, pos, posCodes } from "./position";
+import { makePosition, max, min, pos, posCodes } from "../position";
 
 const mid = (min + max) >> 1;
